[REFACTOR] Update Error Code
에러 수정
